Remove SourceBridge / SourceAPI, and move SourceRuntime to compiler
