+++ amm/compiler/src/main/scala/ammonite/compiler/tools/source.scala
@@ -1,4 +1,4 @@
-package ammonite.repl.tools
+package ammonite.compiler.tools
 
 import ammonite.runtime.tools.browse.Strings
 import ammonite.util.CodeColors
@@ -53,7 +53,7 @@ object source{
                   colors: c.Expr[CodeColors]): c.Expr[Unit] = {
     import c.universe._
     val defaultBrowseExpr = c.Expr[Int => Strings](
-      q"_root_.ammonite.repl.api.SourceBridge.value.browseSourceCommand"
+      q"${prefix(c)}.browseSourceCommand"
     )
 
     applyCustomizeCommandMacro(c)(f, defaultBrowseExpr)(pprinter, colors)
@@ -86,7 +86,7 @@ object source{
 
   def prefix(c: Context) = {
     import c.universe._
-    q"ammonite.repl.api.SourceBridge.value"
+    q"ammonite.compiler.tools.SourceRuntime"
   }
   /**
     * Attempts to up an expression, into either a LHS + methodcall + rhs. We
