+++ amm/interp/src/main/scala/ammonite/interp/Interpreter.scala
@@ -720,7 +720,7 @@ object Interpreter{
     ),
     ImportData("ammonite.runtime.tools.{browse, grep, time}"),
     ImportData("ammonite.runtime.tools.tail", importType = ImportData.TermType),
-    ImportData("ammonite.repl.tools.{desugar, source}"),
+    ImportData("ammonite.compiler.tools.{desugar, source}"),
     ImportData("mainargs.{arg, main}"),
     ImportData("ammonite.repl.tools.Util.PathRead")
   )
