+++ amm/repl/src/test/scala-2.12/ammonite/unit/SourceTests212.scala
@@ -1,18 +1,14 @@
 package ammonite.unit
 
 
-import ammonite.repl.SourceAPIImpl
-import ammonite.repl.api.SourceBridge
 import utest._
-import ammonite.repl.tools.source.load
+import ammonite.compiler.tools.source.load
 import ammonite.util.Util
 import ammonite.util.Util.Location
 //import fastparse.utils.{ElemSetHelper, Generator, IndexedParserInput}
 
 object SourceTests212 extends TestSuite{
   val tests = Tests{
-    if (SourceBridge.value0 == null)
-      SourceBridge.value0 = new SourceAPIImpl {}
     def check(loaded: Location, expectedFileName: String, expected: String, slop: Int = 10) = {
