+++ amm/repl/src/test/scala/ammonite/session/AdvancedTests.scala
@@ -373,7 +373,7 @@ object AdvancedTests extends TestSuite{
     test("desugar"){
       check.session("""
         @ desugar{1 + 2 max 3}
-        res0: ammonite.repl.tools.Desugared = scala.Predef.intWrapper(3).max(3)
+        res0: compiler.tools.Desugared = scala.Predef.intWrapper(3).max(3)
       """)
     }
     test("loadingModulesInPredef"){
