+++ amm/repl/src/test/scala/ammonite/unit/SourceTests.scala
@@ -1,19 +1,14 @@
 package ammonite.unit
 
 
-import ammonite.repl.SourceAPIImpl
-import ammonite.repl.api.SourceBridge
 import utest._
-import ammonite.repl.tools.source.load
+import ammonite.compiler.tools.source.load
 import ammonite.util.Util
 import ammonite.util.Util.Location
 
 import java.io.InputStream
 object SourceTests extends TestSuite{
   val tests = Tests{
-
-    if (SourceBridge.value0 == null)
-      SourceBridge.value0 = new SourceAPIImpl {}
 
     def check(loaded: Location, expectedFileName: String, expected: String, slop: Int = 10) = {
